+++ src/Neutron/TemporaryFilesystem/TemporaryFilesystem.php
@@ -27,13 +27,13 @@ class TemporaryFilesystem implements TemporaryFilesystemInterface
     /**
      * {@inheritdoc}
      */
-    public function createTemporaryDirectory($mode = 0777, $maxTry = 65536)
+    public function createTemporaryDirectory($mode = 0777, $maxTry = 65536, $prefix = null)
     {
         $basePath = sys_get_temp_dir();
 
         while ($maxTry > 0) {
             $dir = $basePath . DIRECTORY_SEPARATOR
-                . base_convert(mt_rand(0x19A100, 0x39AA3FF), 10, 36);
+                . $prefix . base_convert(mt_rand(0x19A100, 0x39AA3FF), 10, 36);
 
             if (false === file_exists($dir)) {
                 try {