+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 
   * Use custom IOException instead of Symfony one.
   * Add a TemporaryFilesystemInterface.
+  * Add optional directory prefix to createTemporaryDirectory.
 
 * 2.0.1 (2013-04-09)